more refactoring

+++ src/app/manage-org-page.component.ts
@@ -10,17 +10,20 @@ import { UIHelper, Utilities } from './services/app.service';
 
 @Component({
 	selector: 'manage-org-page',
-	templateUrl: 'app/manage-org-page.component.html'
+	templateUrl: 'app/manage-org-page.component.html',
+	styleUrls: ['app/manage-org-page.component.css']
 })
 
 // Tell users to go to compressjpeg.com if their images exceed 2 MB
 
 export class ManageOrgPageComponent implements OnInit {
 	@Input() org:any; // Declared as an input in case you're including it inside another component like <manage-org-page [org]="org">...
+	@Input() user;
 	private sub:Subscription;
 	private isLoaded:boolean = false;
 	private stillWorking:boolean = false;
 	private progress:number = 0;
+	private adminToken:string;
 
 	/** Fields to edit **/
 	private coverImage:string;
@@ -38,7 +41,7 @@ export class ManageOrgPageComponent implements OnInit {
 		"Help out",
 		"Volunteer"
 	];
-	private categories = this.categoryService.list();
+	private categories = this.categoryService.list(); 
 
 	/** Saving (most are unused in template but must exist or else save() will break) **/
 	private saving_coverImage:boolean;
@@ -50,6 +53,8 @@ export class ManageOrgPageComponent implements OnInit {
 
 	/** Changed **/
 	private changed_otherLinks:boolean;
+	private changed_categories:boolean;
+	private checked = {};
 
 	/** Slug validation **/
 	private slugIsValid:boolean = true;
@@ -73,6 +78,15 @@ export class ManageOrgPageComponent implements OnInit {
 		this.ui.setTitle("Manage");
 		this.userService.getLoggedInUser((err, user) => {
 			if (err) return console.error(err);
+			this.user = user;
+			this.http.get("/adminToken").map(res => res.json()).subscribe(
+				data => {
+					this.adminToken = data;
+				},
+				err => {
+					console.error(err);
+				}
+			);
 			if (this.route.params) {
 				this.sub = this.route.params.subscribe(params => {
 					let id = params['id'];
@@ -83,7 +97,7 @@ export class ManageOrgPageComponent implements OnInit {
 
 					this.orgService.loadOrg(id).subscribe(
 						data => {
-							if (user.adminToken !== 'h2u81eg7wr3h9uijk8') {
+							if (!this.userIsAdmin()) {
 								if (!data || !data._id || user.permissions.indexOf(data.globalPermission) === -1) {
 									this.ui.flash("Either the page doesn't exist or you don't have permission to manage it", "error");
 									return this.router.navigate([''], { queryParams: {"404": true}});
@@ -91,6 +105,8 @@ export class ManageOrgPageComponent implements OnInit {
 							}
 							this.org = data;
 							this.isLoaded = true;
+							this.org.categories.forEach(category => this.checked[category.id] = true);
+							this.org.description = this.org.description.replace(/(?:\r\n|\r|\n)/g, '<br />');
 							
 							// for ng-upload
 							this.coverImageUploadOptions = {
@@ -142,12 +158,12 @@ export class ManageOrgPageComponent implements OnInit {
   } 
 
   save(key:string, value?:any):void {
-  	if (key === "categories") {
-  		value = this.org.categories;
-  	}
-
   	if (typeof value === "undefined") {
   		value = this[key] || this.org[key];
+  	}
+
+  	if (key === "categories") {
+  		value = this.org.categories;
   	}
 
   	if (key === "slug") {
@@ -181,7 +197,6 @@ export class ManageOrgPageComponent implements OnInit {
   			return;
   		}
   		this.org = res;
-  		this[key] = null;
   		this['saving_' + key] = false;
   		this['changed_' + key] = false;
   		this.ui.flash("Saved", "success");
@@ -199,7 +214,6 @@ export class ManageOrgPageComponent implements OnInit {
   }
 
   changeSelectedCategories(category, add) {
-  	console.log(this.org.categories);
   	let categoryIndex = -1;
   	let foundCategory = this.org.categories.find((cat, index) => {
   		if (cat.id == category.id) {
@@ -212,25 +226,26 @@ export class ManageOrgPageComponent implements OnInit {
   	if (add) {
   		if (categoryIndex === -1) {
   			this.org.categories.push(category);
+  			this.checked[category.id] = true;
   		}
-  		console.log("Added", category);
   	} 
   	else {
 	  	this.org.categories.splice(categoryIndex, 1);
-	  	console.log("Removed", category);
+	  	this.checked[category.id] = false;
 	  }
-	  console.log(this.org.categories);
   }
 
   deleteOrg(id) {
   	if (window.confirm("Are you sure you want to delete this organization? This can't be undone.")) {
-	  	let orgId = id || this.org._id;
-	  	this.http.delete('/org/' + orgId).map(res => res.json()).subscribe(data => {
-	  		if (data && data.success) {
-	  			this.router.navigate(['']);
-	  			return this.ui.flash("Org was deleted", "error");
-	  		}
-	  	});
+  		if (window.confirm("Sure you're sure?")) {
+		  	let orgId = id || this.org._id;
+		  	this.http.delete('/org/' + orgId).map(res => res.json()).subscribe(data => {
+		  		if (data && data.success) {
+		  			this.router.navigate(['']);
+		  			return this.ui.flash("Org was deleted", "error");
+		  		}
+		  	});
+		  }
 	  }
   }
 
@@ -245,6 +260,10 @@ export class ManageOrgPageComponent implements OnInit {
   		this['changed_' + key] = true;
   	else
   		this['changed_' + key] = false;
+  }
+
+  userIsAdmin() {
+  	return this.user.adminToken === this.adminToken;
   }
 
 }
